requirements.txt
astroid==2.4.2
backcall==0.2.0
beautifulsoup4==4.9.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.14
debugpy==1.0.0rc2
decorator==4.4.2
Flask==1.1.2
html5lib==1.0.1
idna==2.8
inflection==0.3.1
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.4.2
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jupyter-client==6.1.5
jupyter-core==4.6.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
more-itertools==5.0.0
multitasking==0.0.9
ndg-httpsclient==0.5.1
numpy==1.19.0
pandas==0.25.3
pandas-datareader==0.8.1
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
pyasn1==0.4.8
pycparser==2.19
Pygments==2.6.1
pylint==2.6.0
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
pywin32==228
pyzmq==19.0.1
Quandl==3.4.8
requests==2.22.0
rope==0.17.0
selenium==3.141.0
six==1.15.0
soupsieve==1.9.5
TA-Lib==0.4.18
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yfinance==0.1.54
